+++ src/featuresToUnlock/unlockableAppIcons.ts
